store_service/pyproject.toml
[tool.poetry]
name = "store-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.8"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
requests = "^2.32.3"
responses = "^0.25.3"
drf-spectacular = "^0.27.2"
celery = "^5.4.0"
redis = "^5.0.8"
pylint = "^3.2.6"
django-cors-headers = "^4.4.0"
unidecode = "^1.3.8"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
